Routing the user to different screens depending on if the address is set
