+++ examples/wallet/app/Routes.js
@@ -4,12 +4,13 @@ import routes from './constants/routes';
 import App from './containers/App';
 import Home from './containers/Home';
 import Index from './containers/Index';
+import InputKeys from './containers/InputKeys';
 
 export default () => (
   <App>
     <Switch>
       <Route path={routes.HOME} component={Home} />
-      <Route path={routes.INPUT_KEYS} component={() => <p> TODO</p>} />
+      <Route path={routes.INPUT_KEYS} component={InputKeys} />
       <Route path={routes.INDEX} component={Index} />
     </Switch>
   </App>
